+++ lib/src/analyzers/lint_analyzer/rules/rules_factory.dart
@@ -7,6 +7,7 @@ import 'rules_list/avoid_non_null_assertion/avoid_non_null_assertion.dart';
 import 'rules_list/avoid_preserve_whitespace_false/avoid_preserve_whitespace_false.dart';
 import 'rules_list/avoid_returning_widgets/avoid_returning_widgets.dart';
 import 'rules_list/avoid_unnecessary_setstate/avoid_unnecessary_setstate.dart';
+import 'rules_list/avoid_unnecessary_type_assertions/avoid_unnecessary_type_assertions.dart';
 import 'rules_list/avoid_unused_parameters/avoid_unused_parameters.dart';
 import 'rules_list/avoid_wrapping_in_padding/avoid_wrapping_in_padding.dart';
 import 'rules_list/binary_expression_operand_order/binary_expression_operand_order.dart';
@@ -50,6 +51,8 @@ final _implementedRules = <String, Rule Function(Map<String, Object>)>{
       AvoidReturningWidgetsRule(config),
   AvoidUnnecessarySetStateRule.ruleId: (config) =>
       AvoidUnnecessarySetStateRule(config),
+  AvoidUnnecessaryTypeAssertions.ruleId: (config) =>
+      AvoidUnnecessaryTypeAssertions(config),
   AvoidUnusedParametersRule.ruleId: (config) =>
       AvoidUnusedParametersRule(config),
   AvoidWrappingInPaddingRule.ruleId: (config) =>